refactoring from stateless to stateful widgets
