--- lib/widgets/user_transaction.dart
@@ -1,40 +1,0 @@
-import 'package:flutter/material.dart';
-
-import '../models/transaction.dart';
-import './new_transaction.dart';
-import './transaction_list.dart';
-
-class UserTransactions extends StatefulWidget {
-  @override
-  State<UserTransactions> createState() => _UserTransactionsState();
-}
-
-class _UserTransactionsState extends State<UserTransactions> {
-  final List<Transaction> _userTransactions = [
-    Transaction(id: 't1', title: 'q', amount: 10, date: DateTime.now()),
-    Transaction(id: 't2', title: 'w', amount: 11.1, date: DateTime.now()),
-  ];
-
-  void _addTransaction(String title, double amount) {
-    final newTransaction = Transaction(
-        id: DateTime.now().toString(),
-        title: title,
-        amount: amount,
-        date: DateTime.now(),
-    );
-
-    setState(() {
-      _userTransactions.add(newTransaction);
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        NewTransaction(_addTransaction),
-        TransactionList(_userTransactions),
-      ],
-    );
-  }
-}
